pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-filter-mcp"
version = "0.1.0"
description = "Local MCP server for running restricted Python text filters over files"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.12,<2.0",
  "pydantic>=2,<3",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[project.scripts]
data-filter-mcp = "data_filter_mcp.server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
packages = ["data_filter_mcp", "data_filter_mcp.loaders"]
